requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
werkzeug==2.3.7
flask-bcrypt==4.0.1
python-dotenv==1.0.0
pyopenssl==23.2.0
markupsafe==2.1.3